+++ src/app/store/actions.ts
@@ -2,8 +2,8 @@ import { createAction } from '@reduxjs/toolkit';
 
 import { City } from '../../entities/city';
 import { OfferMaximum, OfferPreview } from '../../entities/offer-card';
-import { CommentGet } from '../../entities/review/interfaces';
-import { User } from '../../entities/user/interfaces';
+import { CommentGet } from '../../entities/review';
+import { User } from '../../entities/user';
 import { AuthorizationStatus, FetchStatus } from '../consts';
 
 export const setCity = createAction<City>('city/set');
