+++ src/entities/Review/index.ts
@@ -1,2 +1,3 @@
 export { Review } from './ui/review';
 export { ReviewsList } from './ui/reviews-list';
+export type { CommentGet } from './interfaces';
